Remove the RecordingCommand from EEFTextImpl.updateValue()
RecordingCommand implies we want a read/write transaction, which
requires eclusive lock on the ResourceSet and can not be done from a
post-commit listener (which is where we are un updateValue()), causing
EMF Tx to throw an exception.

From an EMF Tx point of view updateValue() only reads the model so
nothing special is needed.

Change-Id: Ie22c21fa5339793bae5b85350560cf1ff33cb5dd
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>